feat: create FileInput component using UploadFile component
