+++ src/components/input-components/file-input/style.scss
