+++ src/components/input-components/index.js
@@ -1,9 +1,11 @@
 import ChoiceInput from './choice-input';
 import TextInput from './text-input';
 import MultiSelectInput from './multi-select-input';
+import FileInput from './file-input';
 
 export {
   ChoiceInput,
   TextInput,
   MultiSelectInput,
+  FileInput,
 };
